Move despawn hacks into own plugin
